Move Provider tests from inject.test.js to Provider.test.js
